requirements.txt
alembic==1.4.3
click==7.1.2
fastapi==0.63.0
h11==0.11.0
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2==2.8.4
pydantic==1.7.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.2
